Start parsing ".json" files (#974)
Starting with this PR, ".json" files for the new "traceroute3"
datatype are no longer ignored.

This PR includes some improvements to the testing code of the
"parser" package.  Specifically, super long lines (some greater
than 6,000 characters!) that are used to initialize a string
variable are replaced by reading the contents of a file.  Also,
temporary files created during package testing are now deleted
when the test finishes.

The changes were tested as follows:
$ git status -s
$ go test ./...
$ git status -s