Fix bug with snapshot thinning.
In the current version, the snapshot thinning does not work properly because the `n%10 != 0` check, intended to always add the final snapshot, suffers from an off-by-one error.

Concretely, the main loop in the function appends snapshots at index 0, 10, 20, 30, ...; i.e. the first element and every tenth after.
Second, the final check tests whether the list of snapshots is a multiple of 10, and if not adds the final element. And this is where the problem lies: the loop does not take every tenth element, it takes the first one and then every tenth _after_, i.e. the 11th, 21st, 31st, etc -- it is essentially shifted by 1, and thus the `n%10` check is incorrect. This leads to the following behavior: If the list has a length that is a multiple of 10, the final element will be excluded. If it has a length of a multiple of 10 _plus 1_, the final element will be duplicated. Here's an illustration:

```
IN:  []
OUT: []
---
IN:  [0]
OUT: [0, 0]
---
IN:  [0, 1]
OUT: [0, 1]
---
IN:  [0, 1, 2]
OUT: [0, 2]
---
IN:  [0, 1, 2, 3]
OUT: [0, 3]
---
IN:  [0, 1, 2, 3, 4]
OUT: [0, 4]
---
IN:  [0, 1, 2, 3, 4, 5]
OUT: [0, 5]
---
IN:  [0, 1, 2, 3, 4, 5, 6]
OUT: [0, 6]
---
IN:  [0, 1, 2, 3, 4, 5, 6, 7]
OUT: [0, 7]
---
IN:  [0, 1, 2, 3, 4, 5, 6, 7, 8]
OUT: [0, 8]
---
IN:  [0, 1, 2, 3, 4, 5, 6, 7, 8, 9]
OUT: [0]
---
IN:  [0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10]
OUT: [0, 10, 10]
---
IN:  [0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11]
OUT: [0, 10, 11]
```

In this PR, I simply the code somewhat to avoid the index check at all: The main loop only considers elements up to, but excluding the last one (`i < n-1`), which is always added below.